Add a top bar with blog and projects link.
Try to add a top navigation bar with links to the homepage (blog) and the
projects page.  The projects sub-url will look for a file named
projects.mdwn in the base directory, and use that for the index
file.  Make that inline all pages under the projects directory.  I
think this really should only be one page, and have further
sub-directories under projects for different categories of projects, like
irrigation, pest control, etc.

Signed-off-by: Sarah Sharp <sarah@thesharps.us>