Attempt to structure project pages.
Ikiwiki wants to display the name of the inlined page at the top of the
inline.  To make that look right for the project page, we don't want to
have the main project's page show up in the inline.  We may want to have
the subproject names show up though, like irrigation.

Signed-off-by: Sarah Sharp <sarah@thesharps.us>